# src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Food FiApp
  version: 1.0.0
  description: API para gerenciar usuarios (dono de restaurante e cliente), restaurantes e itens de cardapio.
servers:
  - url: http://localhost:8080/api/v1
...
paths:
...
  /user-types:
    get:
      operationId: getUserTypes
      tags:
        - user-types
      summary: Lista todos os tipos de usuario
      description: Retorna uma lista de todos os tipos de usuario cadastrados no sistema. A lista pode ser vazia se nao houver tipos de usuario.
      security:
        - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserTypeResponse'
          description: Lista de tipos de usuario retornada com sucesso (pode ser vazia).

    post:
      operationId: createUserType
      tags:
        - user-types
      summary: Cria um novo tipo de usuario
      description: Cadastra um novo tipo de usuario no sistema. Apenas administradores podem realizar esta operacao.
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserTypeRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserTypeResponse'
          description: Tipo de usuario criado com sucesso. Retorna os detalhes do tipo de usuario recem-criado.
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Dados invalidos fornecidos na requisicao.
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Tipo de usuario ja existe ou conflito de dados.
  /user-types/{uuid}:
    get:
      operationId: getUserType
      tags:
        - user-types
      summary: Busca um tipo de usuario pelo UUID
      description: Retorna os detalhes de um tipo de usuario especifico, utilizando seu UUID unico.
      security:
        - bearerAuth: []
      parameters:
        - name: uuid
          schema:
            format: uuid
            type: string
          in: path
          required: true
          description: O UUID unico do tipo de usuario a ser buscado.
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserTypeResponse'
          description: Tipo de usuario encontrado e seus detalhes retornados com sucesso.
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Tipo de usuario nao encontrado para o UUID fornecido.

    put:
      operationId: updateUserType
      tags:
        - user-types
      summary: Atualiza um tipo de usuario existente
      description: Permite a atualizacao de informacoes de um tipo de usuario existente. Apenas administradores podem realizar esta operacao.
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserTypeRequest'
        required: true
      parameters:
        - name: uuid
          schema:
            format: uuid
            type: string
          in: path
          required: true
          description: O UUID unico do tipo de usuario a ser atualizado.
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserTypeResponse'
          description: Tipo de usuario atualizado com sucesso. Retorna os detalhes atualizados do tipo de usuario.
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Tipo de usuario nao encontrado para o UUID fornecido.
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflito de dados ao atualizar o tipo de usuario.

    delete:
      operationId: deleteUserType
      tags:
        - user-types
      security:
        - bearerAuth: []
      summary: Deleta um tipo de usuario existente
      description: Remove um tipo de usuario do sistema com base no seu UUID. Esta operacao so pode ser realizada por administradores e nao e permitida se o tipo de usuario estiver sendo usado por usuarios ativos.
      parameters:
        - name: uuid
          schema:
            format: uuid
            type: string
          in: path
          required: true
          description: O UUID unico do tipo de usuario a ser deletado.
      responses:
        '204':
          description: Tipo de usuario deletado com sucesso. Nenhuma resposta de conteudo.
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Tipo de usuario nao encontrado para o UUID fornecido.
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Nao e possivel deletar o tipo de usuario pois esta sendo usado por usuarios ativos.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Insira o token JWT obtido do endpoint /auth/login no formato: Bearer {token}
  schemas:
    ErrorResponse:
      title: Resposta de Erro
      required:
        - message
        - code
        - status
        - path
        - timestamp
      type: object
      properties:
        message:
          description: Mensagem de erro detalhada.
          type: string
          example: Dados de entrada invalidos.
        code:
          description: Codigo interno do erro.
          type: string
          example: INVALID_INPUT
        status:
          description: Codigo de status HTTP.
          type: integer
          example: 400
        path:
          description: URL da requisicao que gerou o erro.
          type: string
          example: /api/v1/users
        timestamp:
          description: Data e hora quando o erro ocorreu.
          type: string
          format: date-time
          example: '2025-09-23T22:30:00-03:00'
        fieldErrors:
          description: Mapa de erros de validacao especificos do campo, onde a chave e o nome do campo e o valor e a mensagem de erro.
          type: object
          additionalProperties:
            type: string
          example:
            email: O formato do e-mail e invalido.
            password: A senha deve ter no minimo 8 caracteres.
...
    CreateUserTypeRequest:
      title: Solicitacao de Criacao de Tipo de Usuario
      description: Estrutura da requisicao para criar um novo tipo de usuario no sistema.
      required:
        - name
      type: object
      properties:
        name:
          type: string
          description: 'Nome unico do tipo de usuario (ex: CUSTOMER, OWNER, ADMIN).'
          example: CUSTOMER
          minLength: 3
          maxLength: 50
          pattern: '^[A-Z_]+$'
    UpdateUserTypeRequest:
      title: Solicitacao de Atualizacao de Tipo de Usuario
      description: Estrutura da requisicao para atualizar um tipo de usuario existente.
      required:
        - name
      type: object
      properties:
        name:
          type: string
          description: Novo nome para o tipo de usuario (deve ser unico).
          example: CUSTOMER_VIP
          minLength: 3
          maxLength: 50
          pattern: '^[A-Z_]+$'
    UserTypeResponse:
      title: Resposta de Tipo de Usuario
      description: Estrutura da resposta contendo os detalhes de um tipo de usuario.
      type: object
      required:
        - uuid
        - name
      properties:
        uuid:
          type: string
          format: uuid
          description: Identificador unico do tipo de usuario.
          example: 123e4567-e89b-12d3-a456-426614174000
        name:
          type: string
          description: Nome do tipo de usuario.
          example: CUSTOMER
        isActive:
          type: boolean
          description: Indica se o tipo de usuario esta ativo.
          example: true
        createdAt:
          type: string
          format: date-time
          description: Data e hora de criacao do tipo de usuario.
          example: '2025-01-15T14:30:00-03:00'
        updatedAt:
          type: string
          format: date-time
          description: Data e hora da ultima atualizacao do tipo de usuario.
          example: '2025-01-20T10:15:30-03:00'